requirements.txt
Django==1.10.2
django-appconf==1.0.2
django-debug-toolbar==1.6
django-bootstrap3==7.0.1
django-compressor==2.1
django-debug-toolbar==1.6
django-eztables==0.3.2
django-extensions==1.7.5
django-prettyjson==0.2.1
django-widget-tweaks==1.4.1
django.js==0.8.1
djangorestframework==3.4.7
libsass==0.11.2
#mysqlclient==1.3.9
MySQL-python==1.2.5
psycopg2==2.6.2
rcssmin==1.0.6
requests==2.2.1
rjsmin==1.0.12
six==1.10.0
sqlparse==0.2.2
standardjson==0.3.1
